backend/requirements.txt
Django==4.0.1
gunicorn==20.1.0
psycopg2-binary==2.8.5
requests==2.28.1
django-storages==1.11.1
minio==7.1.14
sqlparse==0.2.4
tqdm==4.64.1
db-dtypes==1.1.1
tqdm==4.64.1
pandas==2.0.0
django-pandas==0.6.6
django-material-admin==1.8.6
whitenoise==6.4.0
django-cors-headers==4.1.0
openai==0.27.6
djangorestframework==3.14.0
drf-yasg==1.21.7
newspaper3k==0.2.8
slack_sdk==3.25.0
openai==0.27.6
newspaper3k==0.2.8
beautifulsoup4==4.11.2